+++ localfolder-test.js
@@ -1,4 +1,6 @@
-// cleidigh - update for TB 145.*
+// cleidigh - update for TB 68.*
+
+// v2.0.0 - add option to change store type, empty trash on exit, creation of mail folders
 
 
 // encapsulation objet
@@ -336,6 +338,43 @@ eu.philoux.localfolder.btCreeDossierLocal = async function () {
             return false;
         }
 
+        // test
+
+        var bValid = true;
+        var rep = {};
+        rep.path = dossier;
+
+        try {
+            console.log("LocalFolders: Starting directory READ:", rep.path);
+            let files = await IOUtils.getChildren(rep.path);
+            if (files.length) {
+                console.log("LocalFolders: Directory contains files");
+                files.forEach(file => {
+                    console.log(file)
+                });
+                console.log("LocalFolders: Directory READ passed", rep.path);
+            }
+        } catch (ex) {
+            console.log("LocalFolders: Directory READ failed", rep.path);
+            bValid = false;
+            return bValid;
+        }
+
+        try {
+            console.log("LocalFolders: Starting file WRITE (LFtest.txt):", rep.path);
+            let fname = `${rep.path}\\LFtest.txt`
+            let rv = await IOUtils.writeUTF8(fname, "Test LF write");
+            console.log("LocalFolders: File WRITE passed", fname);
+
+
+        } catch (ex) {
+            console.log("LocalFolders: File WRITE failed", fname);
+
+            bValid = false;
+            return bValid;
+        }
+
+
         // cleidigh - handle storage type, empty trash
         var storeID = document.getElementById("server.storeTypeMenulist").value;
         var emptyTrashOnExit = document.getElementById("server.emptyTrashOnExit").checked;
@@ -640,7 +679,9 @@ var FolderListener = {
     OnItemUnicharPropertyChanged() { },
     OnItemPropertyFlagChanged() { },
     OnItemEvent() { },
-    OnFolderEvent() { },
+    OnFolderEvent() {
+        console.log("evt")
+     },
     
 };
 
@@ -656,6 +697,8 @@ var FolderListener = {
 eu.philoux.localfolder.ValidRepLocal = function (rep) {
     try {
         var bValid = true;
+        return bValid
+
         var item = null;
         var iter = rep.directoryEntries;
         while (iter.hasMoreElements()) {